Move benchmarks to their own module
This avoids all dependencies for users. It also fixes the weird v1
self-dependency.

Fixes #28
