+++ go.mod
@@ -1,7 +1,3 @@
 module github.com/cespare/xxhash/v2
 
-require (
-	github.com/OneOfOne/xxhash v1.2.2
-	github.com/cespare/xxhash v1.1.0 // indirect
-	github.com/spaolacci/murmur3 v0.0.0-20180118202830-f09979ecbc72
-)
+go 1.13
